hw-backend-14/pyproject.toml
[tool.poetry]
name = "hw-backend-14"
version = "0.1.0"
description = ""
authors = ["Baglan Orazbayev <or.baglane@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
django-debug-toolbar = "^5.0.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
